refactor: extract gateway bearer auth into shared middleware factory
